-Organize classes and interfaces in packages
