+++ SmashClone/src/edu/ilstu/uhigh/smashclone/control/InputAdapter.java
@@ -1,4 +1,4 @@
-package edu.ilstu.uhigh.smashclone;
+package edu.ilstu.uhigh.smashclone.control;
 
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
