+++ SmashClone/src/edu/ilstu/uhigh/smashclone/control/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author Basheer
+ *
+ */
+package edu.ilstu.uhigh.smashclone.control;
