+++ SmashClone/src/edu/ilstu/uhigh/smashclone/game/NonInteractable.java
@@ -1,4 +1,4 @@
-package edu.ilstu.uhigh.smashclone;
+package edu.ilstu.uhigh.smashclone.game;
 import java.awt.Graphics;
 
 /* NonInteractable defines an object that does not interact with the
