+++ SmashClone/src/edu/ilstu/uhigh/smashclone/game/PlayerOneKeys.java
@@ -1,4 +1,4 @@
-package edu.ilstu.uhigh.smashclone;
+package edu.ilstu.uhigh.smashclone.game;
 
 import java.awt.event.KeyEvent;
 import java.util.ArrayList;
