+++ SmashClone/src/edu/ilstu/uhigh/smashclone/game/PlayerTwoKeys.java
@@ -1,4 +1,4 @@
-package edu.ilstu.uhigh.smashclone;
+package edu.ilstu.uhigh.smashclone.game;
 
 import java.awt.event.KeyEvent;
 import java.util.ArrayList;
