+++ SmashClone/src/edu/ilstu/uhigh/smashclone/view/Frame.java
@@ -1,4 +1,4 @@
-package edu.ilstu.uhigh.smashclone;
+package edu.ilstu.uhigh.smashclone.view;
 
 import javax.swing.JFrame;
