+++ SmashClone/src/edu/ilstu/uhigh/smashclone/view/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author Basheer
+ *
+ */
+package edu.ilstu.uhigh.smashclone.view;
